+++ src/main/java/ofos/repository/TranslationRepository.java
@@ -8,10 +8,10 @@ import java.util.List;
 
 public interface TranslationRepository extends JpaRepository<TranslationEntity, Long> {
 
-    TranslationEntity findByProductIdAndLang(int id, String lang);
+    TranslationEntity findByProductIdAndLang(long id, String lang);
 
     @Query("SELECT t FROM TranslationEntity t WHERE t.productId IN (SELECT p.ProductID FROM ProvidesEntity p WHERE p.RestaurantID = ?1)" +
             "AND t.lang = ?2")
-    List<TranslationEntity> findTranslationEntitiesByProductIdAndLang(int id, String lang);
+    List<TranslationEntity> findTranslationEntitiesByProductIdAndLang(long id, String lang);
 
 }
